requirements.txt
python-telegram-bot==5.3.0
redis==2.10.5
requests==2.12.4
fakeredis
